- move demonstration event class to example directory
